Move progress bar and detection button to a separate component
